fix(jr): 32-bit Int traps that crash on EVERY watch below watchOS 27
Found on a Series 6 (watchOS 26.6) — the first watch we have ever had that is not
running the 27 beta. Three of four backends crashed immediately and the fourth,
OWRX, crashed the moment its link stuttered.

★★★ On arm64_32 — every watch below watchOS 27, which today is every watch NOT on
the developer beta — Swift's `Int` is THIRTY-TWO BITS, and converting an
out-of-range Double to Int TRAPS.

  ContentView.qualityGlyph
    Int(Date().timeIntervalSinceReferenceDate * 3)   = 2,420,928,000
    Int32.max                                        = 2,147,483,647
  Overflowing since 2023-09-08; before that it fitted. Rendered only while
  `link.linkSettling`, which is why OWRX survived until its first stutter and
  then died — the stutter set the flag, the glyph drew, the app trapped.

  KiwiClient.ts
    Int(Date().timeIntervalSince1970 * 1000)  ≈ 1.78e12 — 800× past Int32.max.
    Constructed with the client, so Kiwi crashed the instant it connected.

  UberClient.sendTune — Int(frequency) in Hz traps past 2.147 GHz, which an
  SDRplay RSP reaches. Hardened, not yet a field failure.

All three are now Int64. Invisible on the Ultra 3 for as long as we have had it:
watchOS 27 is arm64, Int is 64-bit, and none of these can fire.

★★ RELEASE-BLOCKING. Jr 1.0 is in review with this fault, and it is not an
"old watch" problem — it is an arm64_32 problem, so it would crash for almost
everyone until watchOS 27 ships in September.

+++ spike/WristSDR/WristSDR/ContentView.swift
@@ -1404,7 +1404,18 @@ link.setAutoContrast(wfAutoContrast)
     Group {
       if link.linkSettling {
         // Cycle: index advances off the same 12Hz clock the rest of the chrome breathes on.
-        let i = Int(Date().timeIntervalSinceReferenceDate * 3) % settleTints.count
+        // ★★★ Int64, and reduce BEFORE converting. On arm64_32 — every watch below
+        // watchOS 27, which today is every watch not on the beta — Swift's `Int` is
+        // 32 bits, and `timeIntervalSinceReferenceDate * 3` is ~2.42e9, past Int32.max
+        // (2.147e9). Converting an out-of-range Double to Int TRAPS, so this line was a
+        // guaranteed crash the instant the link began settling.
+        // ★ It has been overflowing since 2023-09-08 — before that the product still fitted
+        // in 32 bits, which is why it ever worked. Invisible on the Ultra 3 (watchOS 27,
+        // arm64, 64-bit Int), and that is the only watch we had.
+        // ★★ Symptom trail, 2026-07-29 Series 6: Kiwi died at connect, VibeServer and FM-DX
+        // on or just after connect, and OWRX ran beautifully until one stutter set
+        // linkSettling — then it crashed. Same line every time.
+        let i = Int(Int64(Date().timeIntervalSinceReferenceDate * 3) % Int64(settleTints.count))
         InstanceNodes()
           .stroke(settleTints[i],
                   style: StrokeStyle(lineWidth: 1.6, lineCap: .round, lineJoin: .round))
